requirements.txt
flask
google-genai
python-dotenv
requests
reportlab
pillow
pytesseract
pdfplumber
python-dotenv